revised updates to STATUS cleaning files
